+++ nx/lib/nx/defn/tree.ex
@@ -164,6 +164,7 @@ defmodule Nx.Defn.Tree do
   def apply_args(%T{data: %Expr{op: :io_call, args: args}}, _type, acc, fun) do
     [tensor_expr, callback_spec, template, ref] = args
     {tensor_expr, acc} = Composite.traverse(tensor_expr, acc, fun)
+    {callback_spec, acc} = traverse_token_io_call_spec(callback_spec, acc, fun)
     {[tensor_expr, callback_spec, template, ref], acc}
   end
 
@@ -219,4 +220,11 @@ defmodule Nx.Defn.Tree do
       arg, acc -> {arg, acc}
     end)
   end
+
+  defp traverse_token_io_call_spec({:token_hook, hooked_expr, inner_spec}, acc, fun) do
+    {hooked_expr, acc} = Composite.traverse(hooked_expr, acc, fun)
+    {{:token_hook, hooked_expr, inner_spec}, acc}
+  end
+
+  defp traverse_token_io_call_spec(callback_spec, acc, _fun), do: {callback_spec, acc}
 end